+++ Jiten.Api/Dtos/ExampleSentenceDto.cs
@@ -1,0 +1,12 @@
+using Jiten.Core.Data;
+
+namespace Jiten.Api.Dtos;
+
+public class ExampleSentenceDto
+{
+    public string Text { get; set; }
+    public int WordPosition { get; set; }
+    public int WordLength { get; set; }
+    public Deck? SourceDeckParent { get; set; }
+    public Deck SourceDeck { get; set; }
+}
