+++ Jiten.Web/utils/isTextNsfw.ts
@@ -1,0 +1,4 @@
+export function isTextNsfw(text: string): boolean {
+  const nsfwWords = ['膣壁', '子宮', '中出し', '膣内', '射精', 'セックス', '精液', '亀頭', '肉棒', 'おちんちん', 'おまんこ', '尿道', '液体', 'ちんぽ', 'レイプ', '中だし', 'なかだし', 'オナニー'];
+  return nsfwWords.some((word) => text.includes(word));
+}
